migrate bastion check to VNET auditor
