+++ docs/setup/Setup_Azure.md
@@ -188,11 +188,11 @@ Todo...
 
 ## Azure Checks & Services
 
-These are the following services and checks perform by each Auditor, there are currently **87 Checks** across **6 Auditors** that support the secure configuration of **9 services/components**
+These are the following services and checks perform by each Auditor, there are currently **87 Checks** across **7 Auditors** that support the secure configuration of **9 services/components**
 
 | Auditor File Name | Scanned Resource Name | Auditor Scan Description |
 |---|---|---|
-| Azure_VirtualMachine_Auditor | Azure Virtual Network | VNETs should have Azure Bastion deployed in them |
+| Azure_VNET_Auditor | Azure Virtual Network | VNETs should have Azure Bastion deployed in them |
 | Azure_VirtualMachine_Auditor | Azure Virtual Machine | Azure VMs should use Managed Disks (not VHDs) |
 | Azure_VirtualMachine_Auditor | Azure Virtual Machine | Azure VMs should encrypted OS and Data disks with CMKs |
 | Azure_VirtualMachine_Auditor | Azure Disk Storage Disks | Unattached storage disks should be encrypted with CMKs |
